Classification of Iris Species using the KNN Algorithm
Following is a Basic Classification Program Trained and Tested on the Fisher's Iris Dataset  that contains a set of 150 records of the iris flowers under five attributes - sepal length, sepal width, petal length, petal width and species. The Dataset is divided randomly into two halves into training and testing data on which the program is trained and henceforth tested. For classification and regression purpose, the KNN or the k-Nearest Neighbors Algorithm is used. 
The accuracy of the algorithm is : 0.9736842105263158